Refine aprgaudio through refactoring

+++ AprgAudio/AprgAudio/tst/AudioTests/Audio_unit.cpp
@@ -155,15 +155,13 @@ TEST(WavLoadingTests, Stereo_8bit_44100) {
     EXPECT_EQ(audioFile.getNumSamplesPerChannel(), wav_stereo_8bit_44100::numSamplesPerChannel);
     EXPECT_EQ(audioFile.getBitDepth(), wav_stereo_8bit_44100::bitDepth);
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_8bit_44100::sampleRate);
-    EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_8bit_44100::testBuffer.size());
+    EXPECT_EQ(audioFile.getNumChannels(), static_cast<int>(wav_stereo_8bit_44100::testBuffer.size()));
 
     for (int i = 0; i < static_cast<int>(wav_stereo_8bit_44100::testBuffer[0].size()); i++) {
-        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
-            EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_8bit_44100::testBuffer[k][i]));
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {            EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_8bit_44100::testBuffer[k][i]));
         }
     }
 }
-
 TEST(WavLoadingTests, Stereo_16bit_44100) {
     Audio<double> audioFile;
     bool isLoadingSuccessful =
@@ -173,15 +171,13 @@ TEST(WavLoadingTests, Stereo_16bit_44100) {
     EXPECT_EQ(audioFile.getNumSamplesPerChannel(), wav_stereo_16bit_44100::numSamplesPerChannel);
     EXPECT_EQ(audioFile.getBitDepth(), wav_stereo_16bit_44100::bitDepth);
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_16bit_44100::sampleRate);
-    EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_16bit_44100::testBuffer.size());
+    EXPECT_EQ(audioFile.getNumChannels(), static_cast<int>(wav_stereo_16bit_44100::testBuffer.size()));
 
     for (int i = 0; i < static_cast<int>(wav_stereo_16bit_44100::testBuffer[0].size()); i++) {
-        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
-            EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_16bit_44100::testBuffer[k][i]));
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {            EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_16bit_44100::testBuffer[k][i]));
         }
     }
 }
-
 TEST(WavLoadingTests, Stereo_24bit_44100) {
     Audio<double> audioFile;
     bool isLoadingSuccessful =
@@ -191,15 +187,13 @@ TEST(WavLoadingTests, Stereo_24bit_44100) {
     EXPECT_EQ(audioFile.getNumSamplesPerChannel(), wav_stereo_24bit_44100::numSamplesPerChannel);
     EXPECT_EQ(audioFile.getBitDepth(), wav_stereo_24bit_44100::bitDepth);
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_24bit_44100::sampleRate);
-    EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_24bit_44100::testBuffer.size());
+    EXPECT_EQ(audioFile.getNumChannels(), static_cast<int>(wav_stereo_24bit_44100::testBuffer.size()));
 
     for (int i = 0; i < static_cast<int>(wav_stereo_24bit_44100::testBuffer[0].size()); i++) {
-        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
-            EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_24bit_44100::testBuffer[k][i]));
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {            EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_24bit_44100::testBuffer[k][i]));
         }
     }
 }
-
 TEST(WavLoadingTests, Mono_16bit_44100) {
     Audio<double> audioFile;
     bool isLoadingSuccessful =
@@ -227,15 +221,13 @@ TEST(WavLoadingTests, Stereo_8bit_48000) {
     EXPECT_EQ(audioFile.getNumSamplesPerChannel(), wav_stereo_8bit_48000::numSamplesPerChannel);
     EXPECT_EQ(audioFile.getBitDepth(), wav_stereo_8bit_48000::bitDepth);
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_8bit_48000::sampleRate);
-    EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_8bit_48000::testBuffer.size());
+    EXPECT_EQ(audioFile.getNumChannels(), static_cast<int>(wav_stereo_8bit_48000::testBuffer.size()));
 
     for (int i = 0; i < static_cast<int>(wav_stereo_8bit_48000::testBuffer[0].size()); i++) {
-        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
-            EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_8bit_48000::testBuffer[k][i]));
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {            EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_8bit_48000::testBuffer[k][i]));
         }
     }
 }
-
 TEST(WavLoadingTests, Stereo_16bit_48000) {
     Audio<double> audioFile;
     bool isLoadingSuccessful =
@@ -245,15 +237,13 @@ TEST(WavLoadingTests, Stereo_16bit_48000) {
     EXPECT_EQ(audioFile.getNumSamplesPerChannel(), wav_stereo_16bit_48000::numSamplesPerChannel);
     EXPECT_EQ(audioFile.getBitDepth(), wav_stereo_16bit_48000::bitDepth);
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_16bit_48000::sampleRate);
-    EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_16bit_48000::testBuffer.size());
+    EXPECT_EQ(audioFile.getNumChannels(), static_cast<int>(wav_stereo_16bit_48000::testBuffer.size()));
 
     for (int i = 0; i < static_cast<int>(wav_stereo_16bit_48000::testBuffer[0].size()); i++) {
-        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
-            EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_16bit_48000::testBuffer[k][i]));
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {            EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_16bit_48000::testBuffer[k][i]));
         }
     }
 }
-
 TEST(WavLoadingTests, Stereo_24bit_48000) {
     Audio<double> audioFile;
     bool isLoadingSuccessful =
@@ -263,15 +253,13 @@ TEST(WavLoadingTests, Stereo_24bit_48000) {
     EXPECT_EQ(audioFile.getNumSamplesPerChannel(), wav_stereo_24bit_48000::numSamplesPerChannel);
     EXPECT_EQ(audioFile.getBitDepth(), wav_stereo_24bit_48000::bitDepth);
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_24bit_48000::sampleRate);
-    EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_24bit_48000::testBuffer.size());
+    EXPECT_EQ(audioFile.getNumChannels(), static_cast<int>(wav_stereo_24bit_48000::testBuffer.size()));
 
     for (int i = 0; i < static_cast<int>(wav_stereo_24bit_48000::testBuffer[0].size()); i++) {
-        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
-            EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_24bit_48000::testBuffer[k][i]));
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {            EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_24bit_48000::testBuffer[k][i]));
         }
     }
 }
-
 TEST(WavLoadingTests, Mono_16bit_48000) {
     Audio<double> audioFile;
     bool isLoadingSuccessful =
